requirements.txt
pandas
matplotlib
kafka-python
pyspark
codrone-edu
